change like

+++ frontend/src/utils/api.js
@@ -12,79 +12,79 @@ class Api {
     return Promise.reject(`Что-то тут не так: ${res.status}`);
   }
 
-  getInitialCards(token) {
+  getInitialCards() {
     return fetch(`${this._url}/cards`, {
       method: 'GET',
       headers: this._headers,
       credentials: 'include',
-      Authorization: `Bearer ${token}`,
+      // Authorization: `Bearer ${token}`,
     }).then(this._setError);
   }
 
-  postInitialCards(data, token) {
+  postInitialCards(data) {
     return fetch(`${this._url}/cards`, {
       method: 'POST',
       headers: this._headers,
       credentials: 'include',
-      Authorization: `Bearer ${token}`,
+     // Authorization: `Bearer ${token}`,
       body: JSON.stringify({name: data.name, link: data.link}),
     }).then(this._setError);
   }
 
-  changeAvatar( {avatar}, token ) {
+  changeAvatar( {avatar} ) {
     return fetch(`${this._url}/users/me/avatar`, {
       method: 'PATCH',
       headers: this._headers,
       credentials: 'include',
-      Authorization: `Bearer ${token}`,
+     // Authorization: `Bearer ${token}`,
       body:JSON.stringify({avatar: avatar}),
     }).then(this._setError);
   }
 
-  getInfoUser(token) {
+  getInfoUser() {
     return fetch(`${this._url}/users/me`, {
       method: 'GET',
       headers: this._headers,
       credentials: 'include',
-      Authorization: `Bearer ${token}`,
+     // Authorization: `Bearer ${token}`,
     }).then(this._setError);
   }
 
-  editInfoUser( {name, about}, token ) {
+  editInfoUser( {name, about} ) {
     return fetch(`${this._url}/users/me`, {
       method: 'PATCH',
       headers: this._headers,
       credentials: 'include',
-      Authorization: `Bearer ${token}`,
+      // Authorization: `Bearer ${token}`,
       body: JSON.stringify({name: name, about: about}),
     }).then(this._setError);
   }
 
-  deleteInitialCards(cardId, token) {
+  deleteInitialCards(cardId) {
     return fetch(`${this._url}/cards/${cardId}`, {
       method: 'DELETE',
       headers: this._headers,
       credentials: 'include',
-      Authorization: `Bearer ${token}`,
+      //Authorization: `Bearer ${token}`,
     }).then(this._setError);
   }
 
-  addLike(cardId, token) {
-    return fetch(`${this._url}/cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: this._headers,
-      credentials: 'include',
-      Authorization: `Bearer ${token}`,
-    }).then(this._setError);
-  }
-
-  deleteLike(cardId, token) {
-    return fetch(`${this._url}/cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: this._headers,
-      credentials: 'include',
-      Authorization: `Bearer ${token}`,
-    }).then(this._setError);
+  changeLike(cardId, isLiked) {
+    if (!isLiked) {
+      return fetch(`${this._url}/cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: this._headers,
+        credentials: 'include',
+      // Authorization: `Bearer ${token}`,
+      }).then(this._setError);
+    } else {
+      return fetch(`${this._url}/cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: this._headers,
+        credentials: 'include',
+      // Authorization: `Bearer ${token}`,
+      }).then(this._setError);
+    }
   }
 }
